+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 ChangeLog
 =========
+v1.17.7.1 (2021-01-05)
+-------------------------
+* Fix for push request
+* Fix for not saving client name
+
+
+v1.17.7 (2021-01-01)
+-------------------------
+* Changed POST requests
+
+
+v1.17.6.11 (2020-12-27)
+-------------------------
+* Fixed return value
+
+
+v1.17.6.10 (2020-12-21)
+-------------------------
+* Added function to identify removed bookmarks
+
+
 v1.17.6.9 (2020-12-18)
 -------------------------
 * Fix for Firefox 84